LazyVim/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.38.0
groq==0.11.0
google-generativeai==0.8.0
serpapi==0.1.5
requests==2.32.3
pydantic==2.9.0
python-dotenv==1.0.1
httpx==0.27.2
beautifulsoup4==4.12.3
Pillow==10.4.0
